Labfiles/09-integrate-agent-with-foundry-iq/Python/requirements.txt
azure-ai-projects
azure-identity
python-dotenv
